requirements.txt
cffi==1.15.1
colorama==0.4.5
cryptography==38.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.0
